+++ CHANGELOG.md
@@ -8,7 +8,8 @@ mdr1986x-small-lib changelog
 - Init function for set up callbacks and inner buffer and dispatcher function
 for proccess message and callbacks for byte write, read and message parse
 in simple_exchange
-- EXCH_Write function for send messages in simple_exchange.
+- EXCH_Write function for send messages in simple_exchange
+- Acknowledge callback and parsing of acknowledge bytes in simple_exchange
 
 ### Fixed
 - Using of incorrect function names in queue module